+++ sdk/resourcemanager/mongocluster/armmongocluster/constants.go
@@ -6,7 +6,7 @@ package armmongocluster
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/mongocluster/armmongocluster"
-	moduleVersion = "v0.1.0"
+	moduleVersion = "v0.2.0"
 )
 
 // ActionType - Extensible enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
@@ -48,15 +48,21 @@ type CreateMode string
 const (
 	// CreateModeDefault - Create a new mongo cluster.
 	CreateModeDefault CreateMode = "Default"
+	// CreateModeGeoReplica - Create a replica cluster in distinct geographic region from the source cluster.
+	CreateModeGeoReplica CreateMode = "GeoReplica"
 	// CreateModePointInTimeRestore - Create a mongo cluster from a restore point-in-time.
 	CreateModePointInTimeRestore CreateMode = "PointInTimeRestore"
+	// CreateModeReplica - Create a replica cluster in the same geographic region as the source cluster.
+	CreateModeReplica CreateMode = "Replica"
 )
 
 // PossibleCreateModeValues returns the possible values for the CreateMode const type.
 func PossibleCreateModeValues() []CreateMode {
 	return []CreateMode{
 		CreateModeDefault,
+		CreateModeGeoReplica,
 		CreateModePointInTimeRestore,
+		CreateModeReplica,
 	}
 }
 
@@ -81,39 +87,6 @@ func PossibleCreatedByTypeValues() []CreatedByType {
 		CreatedByTypeKey,
 		CreatedByTypeManagedIdentity,
 		CreatedByTypeUser,
-	}
-}
-
-// MongoClusterStatus - The status of the Mongo cluster resource.
-type MongoClusterStatus string
-
-const (
-	// MongoClusterStatusDropping - The mongo cluster resource is being dropped.
-	MongoClusterStatusDropping MongoClusterStatus = "Dropping"
-	// MongoClusterStatusProvisioning - The mongo cluster resource is being provisioned.
-	MongoClusterStatusProvisioning MongoClusterStatus = "Provisioning"
-	// MongoClusterStatusReady - The mongo cluster resource is ready for use.
-	MongoClusterStatusReady MongoClusterStatus = "Ready"
-	// MongoClusterStatusStarting - The mongo cluster resource is being started.
-	MongoClusterStatusStarting MongoClusterStatus = "Starting"
-	// MongoClusterStatusStopped - The mongo cluster resource is stopped.
-	MongoClusterStatusStopped MongoClusterStatus = "Stopped"
-	// MongoClusterStatusStopping - The mongo cluster resource is being stopped.
-	MongoClusterStatusStopping MongoClusterStatus = "Stopping"
-	// MongoClusterStatusUpdating - The mongo cluster resource is being updated.
-	MongoClusterStatusUpdating MongoClusterStatus = "Updating"
-)
-
-// PossibleMongoClusterStatusValues returns the possible values for the MongoClusterStatus const type.
-func PossibleMongoClusterStatusValues() []MongoClusterStatus {
-	return []MongoClusterStatus{
-		MongoClusterStatusDropping,
-		MongoClusterStatusProvisioning,
-		MongoClusterStatusReady,
-		MongoClusterStatusStarting,
-		MongoClusterStatusStopped,
-		MongoClusterStatusStopping,
-		MongoClusterStatusUpdating,
 	}
 }
 
@@ -151,6 +124,22 @@ func PossibleOriginValues() []Origin {
 		OriginSystem,
 		OriginUser,
 		OriginUserSystem,
+	}
+}
+
+// PreviewFeature - Preview features that can be enabled on a mongo cluster.
+type PreviewFeature string
+
+const (
+	// PreviewFeatureGeoReplicas - Enables geo replicas preview feature. The feature must be set at create-time on new cluster
+	// to enable linking a geo-replica cluster to it.
+	PreviewFeatureGeoReplicas PreviewFeature = "GeoReplicas"
+)
+
+// PossiblePreviewFeatureValues returns the possible values for the PreviewFeature const type.
+func PossiblePreviewFeatureValues() []PreviewFeature {
+	return []PreviewFeature{
+		PreviewFeatureGeoReplicas,
 	}
 }
 
@@ -196,6 +185,38 @@ func PossiblePrivateEndpointServiceConnectionStatusValues() []PrivateEndpointSer
 		PrivateEndpointServiceConnectionStatusApproved,
 		PrivateEndpointServiceConnectionStatusPending,
 		PrivateEndpointServiceConnectionStatusRejected,
+	}
+}
+
+// PromoteMode - The mode to apply to a promote operation.
+type PromoteMode string
+
+const (
+	// PromoteModeSwitchover - Promotion will switch the current replica cluster to the primary role and the original primary
+	// will be switched to a replica role, maintaining the replication link.
+	PromoteModeSwitchover PromoteMode = "Switchover"
+)
+
+// PossiblePromoteModeValues returns the possible values for the PromoteMode const type.
+func PossiblePromoteModeValues() []PromoteMode {
+	return []PromoteMode{
+		PromoteModeSwitchover,
+	}
+}
+
+// PromoteOption - The option to apply to a promote operation.
+type PromoteOption string
+
+const (
+	// PromoteOptionForced - Promote option forces the promotion without waiting for the replica to be caught up to the primary.
+	// This can result in data-loss so should only be used during disaster recovery scenarios.
+	PromoteOptionForced PromoteOption = "Forced"
+)
+
+// PossiblePromoteOptionValues returns the possible values for the PromoteOption const type.
+func PossiblePromoteOptionValues() []PromoteOption {
+	return []PromoteOption{
+		PromoteOptionForced,
 	}
 }
 
@@ -245,5 +266,90 @@ func PossiblePublicNetworkAccessValues() []PublicNetworkAccess {
 	return []PublicNetworkAccess{
 		PublicNetworkAccessDisabled,
 		PublicNetworkAccessEnabled,
+	}
+}
+
+// ReplicationRole - Replication role of the mongo cluster.
+type ReplicationRole string
+
+const (
+	// ReplicationRoleAsyncReplica - The cluster is a local asynchronous replica.
+	ReplicationRoleAsyncReplica ReplicationRole = "AsyncReplica"
+	// ReplicationRoleGeoAsyncReplica - The cluster is a geo-asynchronous replica.
+	ReplicationRoleGeoAsyncReplica ReplicationRole = "GeoAsyncReplica"
+	// ReplicationRolePrimary - The cluster is a primary replica.
+	ReplicationRolePrimary ReplicationRole = "Primary"
+)
+
+// PossibleReplicationRoleValues returns the possible values for the ReplicationRole const type.
+func PossibleReplicationRoleValues() []ReplicationRole {
+	return []ReplicationRole{
+		ReplicationRoleAsyncReplica,
+		ReplicationRoleGeoAsyncReplica,
+		ReplicationRolePrimary,
+	}
+}
+
+// ReplicationState - The state of the replication link between the replica and source cluster.
+type ReplicationState string
+
+const (
+	// ReplicationStateActive - Replication link is active.
+	ReplicationStateActive ReplicationState = "Active"
+	// ReplicationStateBroken - Replication link is broken and the replica may need to be recreated.
+	ReplicationStateBroken ReplicationState = "Broken"
+	// ReplicationStateCatchup - Replica is catching-up with the primary. This can occur after the replica is created or after
+	// a promotion is triggered.
+	ReplicationStateCatchup ReplicationState = "Catchup"
+	// ReplicationStateProvisioning - Replica and replication link to the primary is being created.
+	ReplicationStateProvisioning ReplicationState = "Provisioning"
+	// ReplicationStateReconfiguring - Replication link is re-configuring due to a promotion event.
+	ReplicationStateReconfiguring ReplicationState = "Reconfiguring"
+	// ReplicationStateUpdating - Replication link is being updated due to a change on the replica or an upgrade.
+	ReplicationStateUpdating ReplicationState = "Updating"
+)
+
+// PossibleReplicationStateValues returns the possible values for the ReplicationState const type.
+func PossibleReplicationStateValues() []ReplicationState {
+	return []ReplicationState{
+		ReplicationStateActive,
+		ReplicationStateBroken,
+		ReplicationStateCatchup,
+		ReplicationStateProvisioning,
+		ReplicationStateReconfiguring,
+		ReplicationStateUpdating,
+	}
+}
+
+// Status - The status of the Mongo cluster resource.
+type Status string
+
+const (
+	// StatusDropping - The mongo cluster resource is being dropped.
+	StatusDropping Status = "Dropping"
+	// StatusProvisioning - The mongo cluster resource is being provisioned.
+	StatusProvisioning Status = "Provisioning"
+	// StatusReady - The mongo cluster resource is ready for use.
+	StatusReady Status = "Ready"
+	// StatusStarting - The mongo cluster resource is being started.
+	StatusStarting Status = "Starting"
+	// StatusStopped - The mongo cluster resource is stopped.
+	StatusStopped Status = "Stopped"
+	// StatusStopping - The mongo cluster resource is being stopped.
+	StatusStopping Status = "Stopping"
+	// StatusUpdating - The mongo cluster resource is being updated.
+	StatusUpdating Status = "Updating"
+)
+
+// PossibleStatusValues returns the possible values for the Status const type.
+func PossibleStatusValues() []Status {
+	return []Status{
+		StatusDropping,
+		StatusProvisioning,
+		StatusReady,
+		StatusStarting,
+		StatusStopped,
+		StatusStopping,
+		StatusUpdating,
 	}
 }